backend/pyproject.toml
[project]
name = "formwise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgi-correlation-id>=4.3.4",
    "beanie[srv]>=1.27.0",
    "fastapi[standard]>=0.115.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.1",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "C",  # flake8-comprehensions
    "SIM",  # flake8-simplify
    "I",  # isort
]

[tool.ruff.format]
docstring-code-format = true
